Bump bitvec dependency to 0.22.3
There was a conflict between bitvec and a dependency of it (funty) that
caused a build failure. See:
- https://github.com/bitvecto-rs/bitvec/issues/105
- https://github.com/myrrlyn/funty/issues/3

Due to semver, when importing parity-scale-codec as a library and
resolving bitvec 0.20.1 dependencies, funty 0.12 is pulled, which causes
the aforementioned build failure.  I believe this is not happening when
testing parity-scale-codec itself because Cargo.lock pins funty to
1.1.0, but when importing parity-scale-codec from another crate, this
pinning is sometimes not possible.  Bumping bitvec to 0.22.3 solves this
issue, as it is compatible with funty 0.12.