Pin GH actions to a specific hash
This is useful to prevent supply chain attacks conducted against
upstream actions in use within the CI pipeline. If they're compromised
it can lead to theft of secrets used within the CI pipeline.

To update this in the future, you can make the modification to the
version you want to use and then run `pinact run` to repin the hash.